Upgrade react-router-dom
This updates react-router-dom to not quite the latest version, but
a new enough version that it supports hooks.

This also changes how React Refs are handled in such a way that
it breaks some tests which rely on on reaching through the redux
store into child nodes.  To resolve this, some tests are updated
to use the react-test-renderer instead.

The tests in App.test.jsx were not asserting anything past the
initial toEqual assertion in the path, which is why they are only
now being updated to match links added since the test inception.

Change-Id: Ia80fbfe3cf2d2d275fd8422111ec193c467bf606