Fix change cache corruption on second scheduler start
When a second scheduler starts (actually, this is alos true for the
first but it doesn't matter), it adds watchers for all the changes
in the cache.  When they update, it refreshes its local copy of
those changes.  However, it never primes the cache the first time
the watcher fires.  If it runs the cleanup routine before the first
time it sees an update, it won't have a local record of the cache
item's data UUID, so it will consider it unused and delete it.

Effectively this means that the second scheduler to start is very
likely to delete a large number of in-use cache entries unless
they happen to be updated within the first 5 minutes of startup.

To avoid this, allow the cache to update its entry during the
synthetic watch callback which is generated on startup.  In that
case, the event parameter is None.  By removing the check that
event is not None, the cache will be primed.

Change-Id: Id9269aeda6f75d9cd643181d3357faf9542b7623