Store pipeline status for zuul-web in Zookeeper
After processing a pipeline the scheduler will save a condensed pipeline
status in Zookeeper that is meant to be used by zuul-web. This will save
some Zookeeper I/O as we don't have to refresh the complete pipeline
state in order to generate data for diplay in zuul-web.

This is optimization is especially important as some of the data can be
quite large (repo state/files) and is not displayed at all in the web UI.

Change-Id: I6fd99079d4ece863d9918ce138eb620d61a1486f