Remove pf-c-content CSS class from status page
This is an old PF3 CSS rule which seems to conflict with some of the
PF4 rules. As a result, the styles for the h3 headings (used for the
pipeline names) are styled differently than the other headings used
in zuul-web. Especially the margin is set to 0 which makes the headings
look a little "broken" and the containers for the different pipelines
are crushed together.

This is the only occurrence of this rule, so I just removed it and
adapted the "broken" heading to a PF4 Title. Those are small changes,
but the overall result should look much cleaner and more PF4-like.

Change-Id: I226149588f2f3b6ee5d09aa43e1b8c085390ece1