Implement job endpoints directly in zuul-web
This uses the abide in zuul-web to directly access the job
listing instead of asking the scheduler via RPC.

To allow encoding of MappingProxyTypes within cherrypy's response
handling, this implements a custom cherrypy JSON encoder based on
the ZuulJSONEncoder.

Relevant RPC listener methods are removed as they are not used
anywhere else.

Change-Id: I69d580876cdd290618e5c7b81c5fd0a35488cd12