Fix race condition when updating node requests
There is a race condition when the tree cache listener and
`reviseNodeRequest()` try to update the same node request in parallel.

This can lead to cases where the node request event callback is not
called and no `NodesProvisionedEvent` is generated.

To prevent parallel update we remove the side-effect of updating the
node request from the `lockNodeRequest()` method.

Change-Id: Ie5aae0704e5925a5bcc73cc6bc0bcb91287ab26e