Shard config errors
These can be of unknown quantity and size, so let's shard them
in ZK for safety.

I think that they should never change during the life of a buildset,
but I can't guarantee that with the current code.  To be defensive,
this patch creates a new znode with a unique uuid each time they
are changed.  However, it does nothing to clean up unused znodes
(so we will leak these until the queue item is removed).  We can
inspect logs and determine whether this has happened (a new warning
level log message is added) and determine the best way to proceed.

Change-Id: Ifcb1a64ca3e40005a5659b5ddec4cd45d1aec03b