Clear and load TPCs before updating layouts on secondary schedulers
We observed the following in production: on one scheduler a tenant
reconfiguration event.  On the second, it saw the updated layout
and therefore refreshed its layout.  After that point, changes
which update configuration would fail because of duplicate nodeset
definitions.  It appears that the tpc.parsed_branch_config caches
had duplicate data.  This is likely because the cacheConfig call
operates on the TPCs and the TenantParser expects them to be
empty at the start of processing.

To correct this, add a loadTPCs call in updateTenantLayout so it
more closely matches the existing tenant reconfiguration process.

Also add a test case which demonstrates the error.

Change-Id: I1d42de10ba5163c3834627ba5f91dfcf1ba48f8f