WIP: Fix a bug in getting changed files
The original implementation take into account the changed fils from all commits of a PR.
It causes a bug when files get changed and reverted in those commits.
e.g. A file is added in first commit then removed in second commit, this should should not
be considered as a changed file in the PR.

Change-Id: If3beadb4aee246c17948dc153789f227a37aba2c