WIP: Fix a bug in getting changed files
The original implementation take into account the changed fils from
all commits of a PR.
It causes a bug when files get changed and reverted in those commits.
e.g. A file is added in first commit then removed in second commit,
this file should should not be considered as a changed file in the PR.

Change-Id: I7db8b9d3f3267073c5e1a71f52e75939ffa91773