Fix buildset config_errors
A stray equality instead of assignment operator caused us to treat
the config_errors object as a string when deserializing which
caused exceptions.  Correct that.

Additionally, the pipeline processor calls setConfigurationError
somewhat frequently with the same values.  In these cases, we wrote
the data repeatedly to ZK.  To avoid this, add an equality check
before updating config_errors.  The log warning message is retained
so that we can continue to observe for writes of different values.

In order to implement the equality check, we need to be able to
test for equality of ConfigurationError objects and recursively the
source context and start marks included in the key.  In implementing
this, it was discovered that we are not serializing the snippet in
the start mark, so that is corrected.

Change-Id: Ieb70dc6965bded64db6bd9453e2c641578ef48c7