Use a stable hash for ConfigurationErrorKeys
This hash is used to detect if a two configuration errors are
equivalent.  Since the hash is serialized, we need to make sure
to use a stable hashing algorithm so it matches errors created
on other schedulers.

Hashes must be an integer for python's internal usage though,
so we still have the __hash__ method return a hash() of the
sha256 hex digest.  This should only be used within a single
python process.

Change-Id: Ia1f07759b6d81f96dcce313f4eeab149e202159c