Log null change key deference
It appears that it is somehow possible for us to be unable to resolve
a change from a cache change key:

  File "/usr/local/lib/python3.8/site-packages/zuul/manager/__init__.py", line 192, in resolveChangeKeys
    self._change_cache[change.cache_key] = change
  AttributeError: 'NoneType' object has no attribute 'cache_key'

I do not know how this occurs.  To aid in debugging, add a log line
which will emit the change key that we are unable to resolve.  The
following line will still raise the exception due to the null deref.
I do not yet have a suggestion as to how to handle that, so that
intentionally remains.

Change-Id: I9ddd01e3863d7ec28d12eda9da97bfece1afbaf6