DNM: Upload build requests async
Log analysis shows that a significant time of pipeline processing is
lost while scheduling jobs. One expensive part is pushing the build
request to zk since it's a network operation which includes a
potentially large data due to the repo state. Optimize this by moving
this out of the critical path into a separate thread.

Change-Id: If55548aa04585783617f864caffde15017a9faf0