Prevent duplicate config file entries
It is currently possible to list default zuul config file paths in the
extra-config-paths config directive. Doing so will duplicate the configs
in Zuul which can cause problems. Prevent this entirely via
configuration validation.

Note: The tests currently do not pass due to the early load of config
files by the scheduler as noted via comments in scheduler.py. I suspect
that this is actually a larger bug that should be addressed first, but
I've run out of time to look into that today so am pushing this as is.

Change-Id: I07e9d4d3614cbc6cdee06b2866f7ae41d7779135