Remove QueueItem.debug method
We can end up calling item.debug() repeatedly, and each time would,
if we used updateAttributes or activeContext correctly, write the
data out to ZK.  We don't at the moment, which is lucky.  So the
debug info only gets written to ZK the next time we call
updateAttributes.

However, we should correct this so we do explicitly update the debug
info when it changes, but only once.  Since all of the item.debug
calls are in a single call stack, we remove the debug method and
instead pass in a list which we mutate throughout the call stack.

At the end of the job freezing process, we write out the collected
entries once.

QueueItem.warning is similar, except that it is called much less
frequently and in several widely divergent places in the code,
therefore we take the simple approach and explicitly update the
object each time.

Change-Id: I7385243d1c49bf2809470814736222023500a2e6