Conditionally remove nodeset info
In the olden days, this method would check if there was a nodeset
on the build; if so, remove it.

Nowadays we don't store the nodeset directly; instead we store the
node request and the node info.  We seem to have kept the conditional
around the node request but lost the equivalent for the node info.

Since this can be called many times in the case of a change behind
a dependent failing change not at the head, let's wrap that in
a conditional so we're not constantly throwing exceptions.

We no longer need the ecxeption handler since the only exception that
could be raised was if the info wasn't present.

This is exercised by
tests.unit.test_scheduler.TestScheduler.test_failing_dependent_changes
which doesn't fail because of the exception handling, but the error
can be seen in the logs.

Change-Id: I7e892b7506b12d1046fe7bc715528768d9972263