Add an init phase to scheduler/web startup
This adds a new component state: INITIALIZING and the scheduler and
web components use it when they are creating their initial config.

It is safe for the scheduler to start processing tenant and pipeline
events as soon as it starts because it only processes those for
the tenants that it has already loaded.

However it is not safe for drivers to move events from their
incoming queue into the scheduler since that requires the full
tenant list.  The 4 drivers to which this applies are updated
to wait on config priming.

Zuul-web is already structured to wait until config priming
so does not need a corresponding change.

Change-Id: I36dd4927e583328434e66553aa3ff0cd7469b488