Add an init phase to scheduler/web startup
This adds a new component state: INITIALIZING and the scheduler and
web components use it when they are creating their initial config.

We also add a threading event to the scheduler startup sequence to
make sure that we don't start processing before the initial config
is loaded.  Zuul-web is already structured in that way so does not
need a corresponding change.

Change-Id: I36dd4927e583328434e66553aa3ff0cd7469b488