WIP: Add source attribute to GitConnection
So far, the GitConnection didn't provide the 'source' attribute. This
wasn't a problem as the source attribute was only accessed within the
same connection/driver.

With the ChangeCache we are now accessing the source attribute outside
of the specific connection/driver combination which fails in case of a
GitConnection because the source attribute isn't available. This results
in the following exception:

  Traceback (most recent call last):
    File "/workspace/zuul/zk/vendor/watchers.py", line 177, in _log_func_exception
      result = self._func(data, stat, event)
    File "/workspace/zuul/zk/change_cache.py", line 201, in _cacheItemWatcher
      self._get(key, data_uuid, zstat)
    File "/workspace/zuul/zk/change_cache.py", line 306, in _get
      change = self._changeFromData(data)
    File "/workspace/zuul/zk/change_cache.py", line 410, in _changeFromData
      project = self.connection.source.getProject(change_data["project"])
  AttributeError: 'GitConnection' object has no attribute 'source'

To fix this we add the source attribute which is available on the
GitDriver. This is also what other connections do.

Change-Id: I9503bf117f7fab31038e11eef9db65205d55e2e5