Execute Github tests with only one scheduler
Currently, the canMerge check makes most of the tests fail when they are
executed with multiple schedulers.

The main reason are the fake github classes we are using as "backend" in
our tests. Since each scheduler gets a different (fake) connection, each
connection might have a different set of data during the test depending
on which scheduler handleded the trigger event.

When the canMerge check is executed by the wrong scheduler the test will
fail, as the fake connection bound to this scheduler is not aware of the
commit in question.

As this is mainly a testing issue, we decided to limit the number of
schedulers for the Github tests to 1. As a long term solution we should
implement a HTTP server that serves as fake Github backend (similar to
our FakeGerrit), so that the connection data isn't bound to a single
python object.

Change-Id: Ic4dcab061b5e0a1951e93da22debbb161a1f3ecb