Fix zuul-web startup config priming
We didn't include the branch cache ltimes when loading the initial
config in zuul-web, so if zuul-web starts before the branch cache
is updated, it won't be told to refresh it when loading the layout.

This matches what we do in zuul-scheduler, and adds a skipped test.
The test does unusual things with the startup sequence and does not
appear to be reliable when run upstream, so it is disabled for now.

This also adds a debug line and adjust another to remove an apostraphe
to make grepping easier.

Change-Id: Ib184f4cadb21124fd3dedeea3797b01bb9c431cc