Report more YAML syntax errors
In our override of construct_mapping, we don't handle one of the
error cases that the base class does.  If the user supplies
"foo: {{ bar }}" we fail before the superclass method runs and
therefore we don't benefit from its error.  This change corrects
that by testing for that case and allowing the superclass method
to run.

Change-Id: I0b92053b08cb4991309163c94846bd34c4cb1393