Update patch set 1
Patch Set 1: Code-Review-1

I think this should be rebased on https://review.opendev.org/815207 and then the documentation updated to match, and a release note added.

It should probably also depend on https://review.opendev.org/812410 so that users have insight into what's actually checked out.

This does seem like a slightly better approximation of what should be done than the current system, however, it definitely does not match what will happen for branch-based changes.  So if the goal of this is "make tag-based checkouts more intuitive" then it makes progress toward the goal; if it's "make tag and branch changes behave the same" it does not achieve the goal (it could be argued that it may lead people into a false sense of security in that in some cases it makes tags look a little more like branches, and therefore people will expect them to behave 100% like branches; however, we let the horse out of the barn on that one as soon as we added the containing-branches feature for tags, so that may not be a very good argument against this change).

This may potentially change jobs in hard-to-predict ways.  Jobs that work today may continue to work tomorrow but actually run a playbook from a different branch than what they do today.  I think we need to consider how to deal with that.  Here are some (mutually exclusive) choices:

1) Decide that this is a sufficiently intuitive change that it probably won't cause problems for people because it's probably only going to make things work the way people already think they do.  In this case, we can probably just send an email to zuul-discuss to find out if anyone objects and to warn people about it.

2) Make it an opt-out behavior.  Send an email about it and let folks opt-out with a job attribute if they see it do the wrong thing.

3) Make it opt-in.

4) Make some kind of tool to help predict whether this could potentially affect jobs.  Maybe we could do this using the freeze_job API.  Except I think that only takes branches as input, not tags.  So we may have some enhancement work to do there.

My inclination is to go with either #1 or #4, and I think I'm leaning toward #1.  Either we think this is no big deal and just do it (with consultation and notice), or we're worried about the fallout in which case we should try to help people find and fix the problems.  #2 and #3 rely on people detecting extremely subtle behavior changes, which is why I don't favor them.

Regardless of what we choose here, I think we should consider starting to invest in the framework for #4 so we have tools to deal with similar issues in the future.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Attention: {"person_ident":"Gerrit User 27582 \u003c27582@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}