Increase default SQL connection pool size to 10
When executing the unit tests with multiple schedulers in parallel a lot
of them are failing because there are too many open connections.
This doesn't look like a shutdown problem on Zuul side, but my
assumption is that stestr doesn't free up the resources fast enough when
many tests are executed in parallel.

Thereby I discovered that SQLAlchemy's default pool size is 5 which
might also be a little small when multiple schedulers are running. Thus,
let's try to increase the default pool size to 10 and see if that works
better.

This change also makes the pool size configurable via the
connection/database settings.

Change-Id: Ia2920bf48ff1903698a2cd624a5071b687426e2f