Log and ignore exceptions in pipeline processing
In case we encounter an issue during pipeline processing we don't want
to let that exception bubble up to the run handler exception block as
this might starve out following pipelines in case the issue persists.

Instead we will log the exception at the pipeline level and move on to
the next pipeline.

Change-Id: Ie00444e9094b01b31e91f0660290921b887deb94