Load repo state from pipeline state on executors
Instead of sending the repo state via the build request we only provide
the path to the merge/extra repo state that we already as part of the
pipeline state. The executor will then load the repo states from the
given paths.

This way we reduce the amount of duplicate data we store in Zookeeper
which is especially important for the repo state, as it can become quite
large.

Change-Id: I7dc5d854dbee93af52d25b4462293c85eb7a1a8e