Retry dependency update requests
It is possible for dependency updates requests to fail due to errors
with the source. Previously when this happened the change was ignored by
the pipeline and the user gets no feedback. Chances are high that
reporting back to the source will fail so we can't really notify of this
error.

Instead we retry the requests in the hope that the error is a one off
and we can continue to proceed with the originally requested job work.

Story: 2009687
Change-Id: Id010d8c6809b9f9c012b81992590e54bf5e7e1d8