Update patch set 3
Patch Set 3: Code-Review+2

Note that this is not fully backwards-compatible, but I think we're still at a point in our development that we shouldn't worry too much about that.

This can probably be deployed with a complete shutdown and restart, but without clearing out the ZK data.  I might be wrong about that, and it would be good to plan for an additional restart after clearing ZK if it's necessary.

Patch-set: 3
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2