Update patch set 5
Patch Set 5:

Several of the remaining test failures are due to the data no longer being available to the freeze job api.

This change might have been simpler if we just emulated the current params dict instead of trying to use the job object directly.  Basically, do what you're doing when you construct the params data (remove all the fields that come from the job).  But then on the other side when we read the params from ZK, re-hydrate them with the missing fields from the job object.  Then the executer and web servers could use the same methods and just continue operating on the "params" dictionary like they do today.

However, this is a little more clear and efficient, and most of the work is done already, so we might as well see if we can finish it out.  :)

Patch-set: 5
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>