Don't use RecordingMergeClient.history in TestNonLiveMerges
Currently, the history of the scheduler's merge client is used to assert
the correct number of executed merge jobs. This won't work in a
multi-scheduler setup as we can't say for sure which scheduler submitted
which merge jobs.

Instead, we use the merge_job_history property which gives a condensed
view on the history of all merger APIs that are used within the test.

Change-Id: Ib6240a9548015954024ba4de760b5de361d4428d