Combine different history approaches for merge jobs in tests
Currently, there are two "history approaches" for merge jobs in the unit
tests. Both, the RecordingMergeClient and the HoldableMergerApi store
each submitted job in a local dictionary. Those dictionaries can be used
in tests to assert that a specific number (or type) of merge job was
issued during the test.

The majority of the tests uses the history from the HoldableMergerApi.

This change adapts the two remaining test cases that use the history
from the merge client to use the history from the merger api instead.

Additionally, this change adapts the internal structure of the merger
api's history dictionary to reflect the one from the merge client as
this makes the assertions in the unit tests easier.

Change-Id: I459cbe5001d068a10e9851104e0d7a0e213a07cc