Add SOS documentation and remove spec
This migrates and adapts some of the SOS documentation into the ZK
reference.  It omits some of the detail as that belongs either in
the map below or in the code itself, but it retains the narrative
documentation to help understand the approach.

This removes the SOS spec itself since it is now (approximately)
complete, so we don't have redundant documentation.

Change-Id: I7113f1a14c5e3d68a50b242c0cb05a4eedeb0aec