Add helper method to allow skipping test cases for multi scheduler
This adds a skipIfMultiScheduler function to our test setup that can be
used to decorate test cases (or classes) that should be skipped if the
SCHEDULER_COUNT is greater 1.

Depending on what we want to achieve with our test suite (skip tests to
shorten test runtime or execute as many tests as possible to get a
better feedback during development), it might still be a better idea to
set the scheduler_count to 1 for certain tests. However, this only works
for whole test classes as the scheduler_count is a class variable.

If we only want to skip single test methods, this decorator might be a
better approach.

Change-Id: Ib279bbf0c091dc5df4bfbac14b434f46393af005