Report overall duration when a build set is finished
The overall duration is from a user (developer) point of view, how much
time it takes from the trigger of the build (e.g. a push, a comment,
etc.), till the last build is finished.
This time reflects the user experience of how much time the user needs
to wait.

Change-Id: I81b4ef885835de41c06ea3f4ffddbc60dfd4421e