Report overall duration when a build set is finished
The overall duration is from a user (developer) point of view, how much
time it takes from the trigger of the build (e.g. a push, a comment,
etc.), till the last build is finished.

It takes into account also the time spent in waiting in queue, launching
nodes, preparing the nodes, etc.

Technically it measures between the earliest and latest time we can track
in zuul, i.e. from trigger event creating to report generating.

This time reflects the user experience of how much time the user needs
to wait.

Change-Id: Id4d498959ad69aed7c041652022e1734b4923858