Remove rpcclient use from github tests
It's valuable to make sure that we handle the slightly different
(compared to gerrit) code paths of a github change identifier, so
the existing github RPC client tests are modified to directly use
the event queue system.  This still provides most of the coverage
but doesn't require turning these into full-blown zuul-web tests.

One test also relied on a client dequeue for its action, so the
same approach is used there.

Change-Id: I002ebb19112eeb37a8b9e66b2494848875a0ff7e