Don't delete pipeline summary objects if they have a syntax error
With most ZKObjects, we delete them from ZK if we are unable to
deserialize the JSON data.  The pipeline manager will likely
re-create them if able in that case.

But the pipeline summary object is a unique case.  We read from it
without obtaining a lock, so it's possible (likely even) that a
scheduler is in the middle of writing it out (it's sharded, so it
can be multiple znodes) when a zuul-web reads it.  In that case it
was our intention to ignore the error and use the previous data.

However, since the zkobject base class automatically deletes the
object on error, this could result in deleting the summary from ZK
as it's being written.  In that case we might continue using cached
data (or have no data if we didn't happen to have read from it
already) for an extended period of time until the pipeline updates
again (and that update could have the same problem).

To avoid this, add a class variable to indicate that the pipeline
summary object should not delete corrupt data.  We will assume that
it is in the process of writing (and even if it is legitimately
corrupt, the resolution is the same regardless: wait for the scheduler
to write it again on the next pipeline pass, which it always does).

Change-Id: I6da8e7e01e0a31bf30520fdf9829b2a2f0559c11