Web: fix whitelabel tenants with no auth
In the case that no auth info is provided, we were relying on an
exception handler to catch errors and leave the 'info' field in
redux set to null.  We have such an exception handler for the
non-whitelable case, but we did not have one for whitelabeled
tenants.  Therefore, a whitelabel tenant with no auth info would
raise an exception which broke the app.

To correct this, add an exception handler to the whitelabel case
(configureAuthFromInfo).  Additionally, avoid raising that exception
in the first place (so that we always set auth_params), and then
update the login button accordingly.

Change-Id: I8023bdb0db085de7e5abc664ab1b2f67c0744be4